+++ src/fields/boolean-field/booleanField.stories.tsx
@@ -1,14 +1,5 @@
-import { ReactNode } from "react";
-
-import { BooleanField, booleanField } from "./booleanField";
-import { FieldLabel } from "../../components";
-import { FieldErrors } from "../../components/field-errors";
-import {
-  UseOptionsProps,
-  UseSelectFieldProps,
-  useOptions,
-  useSelectFieldProps,
-} from "../../hooks";
+import { booleanField } from "./booleanField";
+import { RadioGroupField } from "../../components/radio-group/RadioGroupField";
 import { FormStory, fixArgs, meta } from "../../scenarios/StoryForm";
 
 export default {
@@ -16,45 +7,7 @@ export default {
   title: "fields/booleanField",
 };
 
-const YesNoOptions = <Option,>({
-  field,
-  label,
-  getValue,
-  getLabel,
-  options,
-}: { label: ReactNode } & UseOptionsProps<Option> &
-  UseSelectFieldProps<Option, BooleanField>) => {
-  const props = useSelectFieldProps({ field, options, getValue });
-
-  const { renderOptions } = useOptions({
-    field,
-    getLabel,
-    options,
-  });
-
-  return (
-    <div style={{ margin: "20px 0" }}>
-      <FieldLabel field={field} label={label} />
-      {renderOptions.map(({ id, value, label }) => (
-        <div key={id}>
-          <input
-            type="radio"
-            {...props}
-            id={id}
-            value={value}
-            checked={props.value === value}
-          />
-          <label htmlFor={id}>{label}</label>
-        </div>
-      ))}
-      <div>
-        <FieldErrors field={field} />
-      </div>
-    </div>
-  );
-};
-
-const yesNoOptions = [
+const funeralTermsOptions = [
   {
     label: "Yes, I accept the Funeral terms.",
     value: true,
@@ -68,40 +21,54 @@ const yesNoOptions = [
 export const Required: FormStory = {
   args: fixArgs({
     fields: {
-      termsOfService: booleanField(),
+      funeralTerms: booleanField({ name: "funeralTerms" }),
     },
     children: ({ fields }) => (
-      <YesNoOptions
-        field={fields.termsOfService}
+      <RadioGroupField
+        field={fields.funeralTerms}
         label="Do you accept the Funeral terms?"
-        options={yesNoOptions}
+        options={funeralTermsOptions}
         getLabel={({ label }) => label}
         getValue={({ value }) => value}
       />
     ),
   }),
+  parameters: {
+    docs: {
+      description: {
+        story:
+          "Required _booleanField_ requires explicit `true` or `false` value. It starts with `undefined` unlike _checkoxField_ (which starts with `false`) so we know that the user interacted with the field.",
+      },
+    },
+  },
 };
 
-// export const Optional: FormStory = {
-//   args: fixArgs({
-//     fields: {
-//       newsletter: checkboxField({
-//         optional: true,
-//       }),
-//     },
-//     children: ({ fields }) => (
-//       <CheckboxInput
-//         field={fields.newsletter}
-//         label="Subscribe to newsletter"
-//       />
-//     ),
-//   }),
-//   parameters: {
-//     docs: {
-//       description: {
-//         story:
-//           "Optional checkboxField will have true or false value in the submit data.",
-//       },
-//     },
-//   },
-// };
+const surveyOptions = [
+  { label: "I like this experience", key: true },
+  { label: "I don't like this", key: false },
+];
+
+export const Optional: FormStory = {
+  args: fixArgs({
+    fields: {
+      approved: booleanField().optional(),
+    },
+    children: ({ fields }) => (
+      <RadioGroupField
+        field={fields.approved}
+        label="Survey"
+        options={surveyOptions}
+        getValue={({ key }) => key}
+        getLabel={({ label }) => label}
+      />
+    ),
+  }),
+  parameters: {
+    docs: {
+      description: {
+        story:
+          "Optional _booleanField_ can be submitted with the initial `undefined` value which indicates that the user didn't care to interact with the field.",
+      },
+    },
+  },
+};