+++ src/fields/zodField.ts
@@ -1,13 +1,48 @@
-import { FieldAtom, FieldAtomConfig, fieldAtom } from "form-atoms";
+import { FieldAtom, FieldAtomConfig, FormAtom, fieldAtom } from "form-atoms";
 import { zodValidate } from "form-atoms/zod";
 import { Atom, Getter, WritableAtom, atom } from "jotai";
 import { RESET, atomWithReset } from "jotai/utils";
 import { ZodUndefined, z } from "zod";
 
 type ValidationConfig<Schema extends z.Schema, OptSchema extends z.Schema> = {
-  optional?: boolean;
   schema: Schema | ((get: Getter) => Schema);
   optionalSchema?: OptSchema | ((get: Getter) => OptSchema);
+};
+
+type Flatten<T> = Identity<{
+  [K in keyof T]: T[K];
+}>;
+type Identity<T> = T;
+
+export type FormSubmitValues<Form extends FormAtom<any>> =
+  Form extends FormAtom<infer Fields> ? FormFieldSubmitValues<Fields> : never;
+
+type FormFieldSubmitValues<Fields extends FormFields> = Flatten<{
+  [Key in keyof Fields]: Fields[Key] extends ZodField<any>
+    ? ZodFieldSubmitValue<Fields[Key]>
+    : Fields[Key] extends FieldAtom<infer Value>
+    ? Value
+    : Fields[Key] extends FormFields
+    ? FormFieldSubmitValues<Fields[Key]>
+    : Fields[Key] extends Array<infer Item>
+    ? Item extends ZodField<any>
+      ? ZodFieldSubmitValue<Fields[Key]>[]
+      : Item extends FieldAtom<infer Value>
+      ? Value[]
+      : Item extends FormFields
+      ? FormFieldSubmitValues<Item>[]
+      : never
+    : never;
+}>;
+
+type FormFields = {
+  [key: string | number]:
+    | FieldAtom<any>
+    | ZodField<any>
+    | FormFields
+    | FormFields[]
+    | FieldAtom<any>[]
+    | ZodField<any>[];
 };
 
 export type ZodFieldConfig<
@@ -16,55 +51,62 @@ export type ZodFieldConfig<
 > = FieldAtomConfig<Schema["_output"] | OptSchema["_output"]> &
   ValidationConfig<Schema, OptSchema>;
 
-export type ZodFieldValue<Field> = Field extends ZodField<
-  infer _,
-  infer __,
-  infer Value
->
+export type ZodFieldValue<Field> = Field extends FieldAtom<infer Value>
   ? Value
   : never;
+
+type WritableRequiredAtom = WritableAtom<
+  boolean,
+  [boolean | typeof RESET | ((prev: boolean) => boolean)],
+  void
+>;
+
+type ZodFieldSubmitValue<Field> = Field extends ZodField<
+  infer Schema,
+  infer OptSchema,
+  infer Required
+>
+  ? Required extends WritableRequiredAtom
+    ? Schema["_output"] | OptSchema["_output"]
+    : Required extends Atom<boolean>
+    ? Schema["_output"]
+    : never
+  : never;
+
+export type OptionalZodField<
+  Schema extends z.Schema,
+  OptSchema extends z.Schema = ZodUndefined
+> = ZodField<Schema, OptSchema, WritableRequiredAtom>; // for OptionalZodField we can write false to the required atom
 
 export type ZodField<
   Schema extends z.Schema,
   OptSchema extends z.Schema = ZodUndefined,
-  Value = Schema["_output"] | OptSchema["_output"]
-> = FieldAtom<Value> extends Atom<infer R>
-  ? Atom<
+  RequiredAtom = Atom<boolean> // required field have read-only RequiredAtom
+> = FieldAtom<Schema["_output"] | OptSchema["_output"]> extends Atom<infer R>
+  ? {
+      optional: () => OptionalZodField<Schema, OptSchema>;
+    } & Atom<
       R & {
-        required: WritableAtom<
-          boolean,
-          [boolean | typeof RESET | ((prev: boolean) => boolean)],
-          void
-        >;
+        required: RequiredAtom;
       }
     >
   : never;
 
 export const zodField = <
   Schema extends z.Schema,
-  OptSchema extends z.Schema = ZodUndefined,
-  Value = Schema["_output"] | OptSchema["_output"]
+  OptSchema extends z.Schema = ZodUndefined
 >({
-  optional = false, // all fields required similarly as zod is required by default
   schema,
   optionalSchema,
   ...config
 }: ZodFieldConfig<Schema, OptSchema>): ZodField<Schema, OptSchema> => {
-  const requiredAtom = atomWithReset(!optional);
+  const requiredAtom = atom(true); // constant, unwritable when .optional() is not called
 
   const baseFieldAtom = fieldAtom({
-    validate: zodValidate<Value>(
+    validate: zodValidate(
       (get) => {
-        const required = get(requiredAtom);
         const schemaObj = typeof schema === "function" ? schema(get) : schema;
-        const optionalSchemaObj =
-          typeof optionalSchema === "function"
-            ? optionalSchema(get)
-            : optionalSchema;
-
-        const optSchema = optionalSchemaObj ?? schemaObj.optional();
-
-        return required ? schemaObj : optSchema;
+        return schemaObj;
       },
       {
         on: "blur",
@@ -95,5 +137,54 @@ export const zodField = <
 
   zodField.debugLabel = `field/zodField/${config.name ?? zodField}`;
 
-  return zodField;
+  const makeOptional = () => {
+    const requiredAtom = atomWithReset(false);
+    const validateCallback = zodValidate(
+      (get) => {
+        const schemaObj = typeof schema === "function" ? schema(get) : schema;
+
+        const optionalSchemaObj =
+          typeof optionalSchema === "function"
+            ? optionalSchema(get)
+            : optionalSchema;
+
+        const optSchema = optionalSchemaObj ?? schemaObj.optional();
+
+        return optSchema;
+      },
+      {
+        on: "blur",
+        when: "dirty",
+      }
+    ).or({ on: "change", when: "touched" });
+
+    const zodField = atom((get) => {
+      const baseField = get(baseFieldAtom);
+
+      const fieldAtoms = {
+        required: requiredAtom,
+        validate: validateCallback,
+      };
+
+      if (
+        typeof process !== "undefined" &&
+        process.env.NODE_ENV !== "production"
+      ) {
+        Object.entries(fieldAtoms).map(([atomName, atom]) => {
+          atom.debugLabel = `field/${atomName}/${config.name ?? zodField}`;
+        });
+      }
+
+      return {
+        ...baseField,
+        ...fieldAtoms,
+        _validateCallback: validateCallback,
+      };
+    });
+
+    // TODO: no need to call .optional().optional()
+    return { ...zodField, optional: makeOptional };
+  };
+
+  return { ...zodField, optional: makeOptional };
 };