+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# [3.0.0-next.2](https://github.com/MiroslavPetrik/form-atoms-field/compare/v3.0.0-next.1...v3.0.0-next.2) (2023-05-17)
+
+
+### Bug Fixes
+
+* revert checkbox field props ([09a4d80](https://github.com/MiroslavPetrik/form-atoms-field/commit/09a4d80bee4d09b6b5f2f4068ec5046a341d4850))
+
+# [3.0.0-next.1](https://github.com/MiroslavPetrik/form-atoms-field/compare/v2.5.14...v3.0.0-next.1) (2023-05-17)
+
+
+* BREAKING CHANGE: optional field via .optional() method ([de99234](https://github.com/MiroslavPetrik/form-atoms-field/commit/de99234ec45d9cdc53b4750b61abaa8531eb7012))
+
+
+### BREAKING CHANGES
+
+* switch submit value based on required/optional g param
+
+closes 29
+
 ## [2.5.14](https://github.com/MiroslavPetrik/form-atoms-field/compare/v2.5.13...v2.5.14) (2023-05-17)